money actions in separate controller
